+++ cypress/integration/2.4-ConfigureAffiliationWithPayerAuthDisable.spec.js
@@ -1,4 +1,3 @@
-import { syncCheckoutUICustom } from '../support/common/testcase.js'
 import { setWorkspaceInAffiliation } from '../support/affiliation.js'
 import { loginViaCookies } from '../support/common/support.js'
 
@@ -7,9 +6,8 @@ const config = Cypress.env()
 // Constants
 const { name } = config.workspace
 
-describe('Set up affiliation and sync checkout ui custom', () => {
+describe('Set up affiliation with payer Auth as disabled', () => {
   loginViaCookies()
 
-  setWorkspaceInAffiliation(name)
-  syncCheckoutUICustom()
+  setWorkspaceInAffiliation(name, false)
 })