+++ cypress/integration/2.6-discountProduct.spec.js
@@ -1,14 +1,14 @@
 import { loginViaCookies, updateRetry } from '../support/common/support.js'
-import { discountShipping } from '../support/outputvalidation.js'
 import selectors from '../support/common/selectors.js'
+import { discountProduct } from '../support/outputvalidation'
 import { getTestVariables } from '../support/utils.js'
 import { paymentTestCases, orderTaxAPITestCase } from '../support/testcase.js'
 
-describe('Discount Shipping Testcase', () => {
+describe('Discount Product Testcase', () => {
   loginViaCookies()
 
-  const { prefix, productName, tax, totalAmount, postalCode, env } =
-    discountShipping
+  const { prefix, productName, env, tax, totalAmount, postalCode } =
+    discountProduct
 
   // Verify tax via order tax api
   orderTaxAPITestCase(prefix, tax)
@@ -23,24 +23,28 @@ describe('Discount Shipping Testcase', () => {
   it('Updating product quantity to 1', updateRetry(3), () => {
     cy.checkForTaxErrors()
     // Update Product quantity to 1
-    cy.updateProductQuantity(discountShipping, { quantity: '1' })
+    cy.updateProductQuantity(discountProduct, {
+      quantity: '1',
+    })
   })
 
-  it('Updating Shipping Information', updateRetry(3), () => {
+  it('Updating Shipping Information', updateRetry(4), () => {
     cy.checkForTaxErrors()
     // Update Shipping Section
     cy.updateShippingInformation({ postalCode })
   })
 
-  it('Verify tax and total', updateRetry(3), () => {
+  it('Verifying tax and total amounts,discount for a discounted product', () => {
     // Verify Tax
     cy.get(selectors.TaxAmtLabel).last().should('have.text', tax)
     // Verify Total
     cy.verifyTotal(totalAmount)
+    // Verify Discounts
+    cy.get(selectors.Discounts).last().should('be.visible')
   })
 
   paymentTestCases(
-    discountShipping,
+    discountProduct,
     { prefix, approved: true },
     { ...getTestVariables(prefix), orderIdEnv: env }
   )